Add files via upload
I've more less completed the expectimax algorithm but I believe it is not taking the most efficient route. Will be still working on it later to improve its evaluation function or maybe add a heuristic function. Also updated the main driver file to be able to play the game. It is set up to have to command the ai to go one action at a time by the user command